제현T 수업 3차시

+++ src/main/java/com/example/spring/blog/BlogRepository.java
@@ -1,5 +1,6 @@
 package com.example.spring.blog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlogRepository extends JpaRepository<Blog, Integer> {
+public interface BlogRepository extends JpaRepository<Blog, Long> {
+
 }
